working-example/requirements.txt
requests>=2.31.0
streamlit>=1.28.1
pytest>=7.4.3
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.17.0 